Move to common the code to generate command line doc
