+++ mithril-client-cli/src/utils/mod.rs
@@ -3,14 +3,12 @@
 
 mod expander;
 mod feedback_receiver;
-mod markdown_formatter;
 mod progress_reporter;
 mod snapshot;
 mod unpacker;
 
 pub use expander::*;
 pub use feedback_receiver::*;
-pub use markdown_formatter::*;
 pub use progress_reporter::*;
 pub use snapshot::*;
 pub use unpacker::*;
