+++ mithril-common/src/generate_doc/markdown_formatter.rs
