+++ mithril-common/src/generate_doc/mod.rs
@@ -1,11 +1,12 @@
+//! Commands to generate a markdown documentation for the command line.
+
+mod markdown_formatter;
 
 use std::fs::File;
 use std::io::Write;
-
 use clap::{Parser, Command};
 
-use crate::utils;
-use mithril_client::MithrilResult;
+use crate::StdResult;
 
 /// Generate documentation
 #[derive(Parser, Debug, Clone)]
@@ -15,8 +16,9 @@ pub struct GenerateDocCommands {
     output: String,
 }
 impl GenerateDocCommands {
-    pub async fn execute(&self, cmd_to_document: &mut Command) -> MithrilResult<()> {
-        let doc = utils::doc_markdown(cmd_to_document);
+    /// Generate the command line documentation.
+    pub async fn execute(&self, cmd_to_document: &mut Command) -> StdResult<()> {
+        let doc = markdown_formatter::doc_markdown(cmd_to_document);
         
         let mut buffer: File = File::create(&self.output)?;
         buffer.write(b"Generated doc\n\n")?;
